+++ .github/workflows/docs/conf.py
@@ -1,3 +1,12 @@
+import re
+from typing import Any, Dict, Optional, Tuple
+from urllib.parse import urljoin
+
+from docutils import nodes
+from docutils.nodes import Element, TextElement
+from sphinx.application import Sphinx
+from sphinx.environment import BuildEnvironment
+
 # -*- coding: utf-8 -*-
 
 # Configuration file for the Sphinx documentation builder.
@@ -10,9 +19,26 @@ extensions = [
     "sphinx.ext.autosummary",
     "sphinx.ext.intersphinx",
     "sphinx.ext.mathjax",
+    "sphinx_copybutton",
 ]
 
-html_theme = "sphinx_rtd_theme"
+pygments_style = "borland"
+
+html_theme = "sphinx_book_theme"
+
+html_theme_options = {
+    "repository_url": "https://github.com/CQCL/pytket-aqt",
+    "use_repository_button": True,
+    "use_issues_button": True,
+    "logo": {
+        "image_light": "Quantinuum_logo_black.png",
+        "image_dark": "Quantinuum_logo_white.png",
+    },
+}
+
+html_static_path = ["_static"]
+
+html_css_files = ["custom.css"]
 
 # -- Extension configuration -------------------------------------------------
 
@@ -29,13 +55,6 @@ autodoc_member_order = "groupwise"
 
 # The following code is for resolving broken hyperlinks in the doc.
 
-from sphinx.application import Sphinx
-from docutils import nodes
-from docutils.nodes import Element, TextElement
-from sphinx.environment import BuildEnvironment
-from urllib.parse import urljoin
-import re
-from typing import Any, Dict, List, Optional
 
 # Mappings for broken hyperlinks that intersphinx cannot resolve
 external_url_mapping = {
@@ -74,7 +93,8 @@ custom_internal_mapping = {
 def add_reference(
     app: Sphinx, env: BuildEnvironment, node: Element, contnode: TextElement
 ) -> Optional[nodes.reference]:
-    # Fix references in docstrings that are inherited from the base pytket.backends.Backend class.
+    # Fix references in docstrings that are inherited from the
+    # base pytket.backends.Backend class.
     mapping = app.config.external_url_mapping
     if node.astext() in mapping:
         newnode = nodes.reference(
@@ -97,7 +117,7 @@ def correct_signature(
     options: Dict,
     signature: str,
     return_annotation: str,
-) -> (str, str):
+) -> Tuple[str, str]:
 
     new_signature = signature
     new_return_annotation = return_annotation
@@ -106,7 +126,8 @@ def correct_signature(
             new_signature = new_signature.replace(k, v)
         if return_annotation is not None:
             new_return_annotation = new_return_annotation.replace(k, v)
-    # e.g. Replace <CXConfigType.Snake: 0> by CXConfigType.Snake to avoid silent failure in later stages.
+    # e.g. Replace <CXConfigType.Snake: 0> by
+    # CXConfigType.Snake to avoid silent failure in later stages.
     if new_signature is not None:
         enums_signature = re.findall(r"<.+?\: \d+>", new_signature)
         for e in enums_signature:
@@ -120,7 +141,7 @@ def correct_signature(
     return new_signature, new_return_annotation
 
 
-def setup(app):
+def setup(app: Sphinx) -> None:
     app.add_config_value("custom_internal_mapping", {}, "env")
     app.add_config_value("external_url_mapping", {}, "env")
     app.connect("missing-reference", add_reference)