+++ docs/changelog.rst
@@ -1,6 +1,15 @@
 Changelog
 ~~~~~~~~~
 
+0.26.0 (April 2023)
+-------------------
+
+* Change default optimisation level to 2
+* Add Rz gate to aqt backend
+* Add experimental support for multi zone ion trap architectures
+* Dropped support for python 3.8 and 3.9, added support for 3.11
+* Updated pytket version requirement to 1.14.
+
 0.25.0 (November 2022)
 ----------------------